Fix out-of-bounds read on the Aztec special char table
This didn't take into account the possibility of being accessed while
being in Binary mode, but this can actually happen. For the most part
this is hidden though by then reading the NoMode entry of the next row,
which happens to contain the correct value in this case. However, if the
looked up symbol is a dot and the following mode is Binary, we read one
entry behind the table.